Fix CI: lint against the ruff version CI actually runs
My .venv had ruff 0.15.16 while CI pinned 0.16.7. Ruff promotes rules into its
default set between releases, so `ruff check .` passed locally on every commit
this session while CI found 36 errors in code nobody had touched.

Root cause first: ruff and pytest now live in requirements-dev.txt, which CI
installs verbatim, so a local checkout cannot silently drift again.

Then the 36, by kind:

- C408 (23) and RUF046 (6) rewritten mechanically. I first assumed RUF046 was a
  false positive, since round() on a numpy scalar looked like it would return
  np.float32 and need the int() cast -- it returns a Python int, so the casts
  really were dead.
- ISC004 (3): implicit string concatenation inside a list literal in
  validate_ut.report, where a dropped comma would silently merge two output rows
  rather than fail. Parenthesised.
- RUF059: unused T unpacked from g.valid.shape.
- BLE001: the blind except in the batch loop is deliberate -- an unattended
  multi-hour run must not lose completed work to one corrupt file -- so it is a
  documented noqa, and the failure is still recorded in the ledger.

Verified beyond the 220 tests: validate_ut and replay produce byte-identical
output on seq1.

+++ src/early_warning/anticipation_labels.py
@@ -81,9 +81,9 @@ def make_window_labels(
     min_context_s: float = 2.0,
 ) -> list[WindowLabel]:
     """One label per window end, every `stride_s`, after `min_context_s` of history."""
-    n_bins = int(round(horizon_s / bin_s))
-    stride = max(1, int(round(stride_s * ev.fps)))
-    start = int(round(min_context_s * ev.fps))
+    n_bins = round(horizon_s / bin_s)
+    stride = max(1, round(stride_s * ev.fps))
+    start = round(min_context_s * ev.fps)
     onsets = np.array(sorted(s for s, _ in ev.assault), dtype=np.int64)
     labels = []
     for t in range(start, ev.n_frames, stride):
